+++ apps/mac/Sources/Client/ClientChatView.swift
@@ -1084,13 +1084,7 @@ struct ClientChatThread: View {
         showNewest()
         follow.jump()
         followPulse += 1
-        Task {
-            if await model.send(text) {
-                model.clearDraft()
-            } else {
-                model.returnDraft(text)
-            }
-        }
+        Task { await model.sendFromComposer(text) }
     }
 
     private var dropExperienceVisible: Bool {